backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-decouple==3.8
Pillow==10.4.0
djangorestframework-simplejwt==5.3.1
